+++ components/msg/lib.rs
@@ -12,3 +12,4 @@ extern crate malloc_size_of;
 extern crate malloc_size_of_derive;
 
 pub mod constellation_msg;
+pub mod shared_ipc_router;
