+++ components/net_traits/lib.rs
@@ -29,7 +29,8 @@ use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use ipc_channel::router::ROUTER;
 use ipc_channel::Error as IpcError;
 use mime::Mime;
-use msg::constellation_msg::{HistoryStateId, IpcHandle};
+use msg::constellation_msg::HistoryStateId;
+use msg::shared_ipc_router::IpcHandle;
 use servo_url::ServoUrl;
 use std::error::Error;
 use time::precise_time_ns;
