+++ components/script/lib.rs
@@ -62,7 +62,6 @@ mod layout_image;
 mod mem;
 mod microtask;
 mod network_listener;
-pub mod script_ipc_router;
 pub mod script_runtime;
 #[allow(unsafe_code)]
 pub mod script_thread;
